src/requirements.txt
lxml>=4.6.3
pyserial>=3.5
PyQt5>=5.15.4
chardet>=4.0.0
PyDapi2>=0.3.10